pyproject.toml
[tool.poetry]
name = "writerassistant"
version = "0.1.0"
description = ""
authors = ["Bear <csaba.ferenclaszlo@zapier.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.3.1"
Flask-Bootstrap = "^3.3.7.1"
openai = "^0.27.4"
langchain = "^0.0.149"
flask-wtf = "^1.1.1"
tinydb = "^4.7.1"
pinecone-client = "^2.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
